SCRUM-151 refactor: 첨부파일 관리 부분 공통 컴포넌트 화
